monodir: fixing the int label metadata

+++ software/em/xmipp/libraries/reconstruction/resolution_directional.cpp
@@ -1224,8 +1224,8 @@ void ProgResDir::radialAzimuthalResolution(Matrix2D<double> &resolutionMat,
 	for (size_t ii = 0; ii<xrows; ++ii)
 	{
 		objId = mdprefDirs.addObject();
-		int con;
-		con = VEC_ELEM(PrefferredDirHist,ii);
+		size_t con;
+		con = (size_t) VEC_ELEM(PrefferredDirHist,ii);
 		std::cout << ii << " " << con << std::endl;
 		mdprefDirs.setValue(MDL_IDX, ii, objId);
 		mdprefDirs.setValue(MDL_COUNT, con, objId);
